fix(ffi-types): Guid type definition;

+++ crates/ffi-types/src/common.rs
@@ -26,8 +26,8 @@ pub type Bool = i32;
 #[repr(C)]
 pub struct Guid {
     pub data1: u32,
-    pub data2: u32,
-    pub data3: u32,
+    pub data2: u16,
+    pub data3: u16,
     pub data4: [u8; 8],
 }
 pub type LpCGuid = *const Guid;
